phase_1/project_1_outlier_slide_gen/requirements.txt
pandas
numpy
scikit-learn
matplotlib
python-pptx
streamlit
Pillow
